# Optimize streaming decoder by BCE