```
benchmark                                                     old ns/op     new ns/op     delta
Benchmark_Decode_SmallStruct_Stream_GoJson-16                 736           677           -8.02%
Benchmark_Decode_MediumStruct_Stream_GoJson-16                4094          3184          -22.23%
Benchmark_Decode_LargeStruct_Stream_GoJson-16                 56639         43953         -22.40%
```